migrate unused definition validator to 2and3
