+++ src/plugins/validate-semantic/validators/2and3/refs.js
@@ -1,4 +1,5 @@
 import get from "lodash/get"
+import { escapeJsonPointerToken } from "../../../refs-util"
 
 export const validate2And3RefHasNoSiblings = () => system => {
   return system.validateSelectors.all$refs()
@@ -23,4 +24,31 @@ export const validate2And3RefHasNoSiblings = () => system => {
         return acc
       }, [])
     })
+}
+
+// Add warnings for unused definitions
+export const validate2And3UnusedDefinitions = () => (system) => {
+  return system.validateSelectors.all$refs()
+  .then((nodes) => {
+    const references = nodes.map(node => node.node)
+    const errors = []
+    const basePath = system.specSelectors.isOAS3() ?
+      ["components", "schemas"] :
+      ["definitions"]
+
+    system.specSelectors.definitions()
+    .forEach((val, key) => {
+      const escapedKey = escapeJsonPointerToken(key)
+      if(references.indexOf(`#/${basePath.join("/")}/${escapedKey}`) < 0) {
+        const path = [...basePath, key]
+        errors.push({
+          level: "warning",
+          path,
+          message: "Definition was declared but never used in document"
+        })
+      }
+    })
+
+    return errors
+  })
 }
