+++ src/plugins/validate-semantic/validators/refs.js
@@ -1,30 +1,3 @@
-import get from "lodash/get"
-import { escapeJsonPointerToken } from "../../refs-util"
-
-// Add warnings for unused definitions
-export const validateUnusedDefinitions = () => (system) => {
-  return system.validateSelectors.all$refs()
-  .then((nodes) => {
-    const references = nodes.map(node => node.node)
-    const errors = []
-
-    system.specSelectors.definitions()
-    .forEach((val, key) => {
-      key = escapeJsonPointerToken(key)
-      if(references.indexOf(`#/definitions/${key}`) < 0) {
-        const path = ["definitions", key]
-        errors.push({
-          level: "warning",
-          path,
-          message: "Definition was declared but never used in document"
-        })
-      }
-    })
-
-    return errors
-  })
-}
-
 export const validateRefPathFormatting = () => (system) => {
   return system.validateSelectors.all$refs()
   .then((refArtifacts) => {
