+++ test/plugins/validate-semantic/refs.js
@@ -4,118 +4,6 @@ import validateHelper, { expectNoErrorsOrWarnings } from "./validate-helper.js"
 describe("validation plugin - semantic - refs", function() {
   this.timeout(10 * 1000)
 
-  describe("Unused definitions", () => {
-
-    it("should return a warning when a definition is declared but not used", () => {
-      const spec = {
-        paths: {
-          "/CoolPath": {}
-        },
-        definitions: {
-          abc: {
-            type: "string"
-          }
-        }
-      }
-
-      return validateHelper(spec)
-      .then(system => {
-        const allErrors = system.errSelectors.allErrors().toJS()
-        expect(allErrors.length).toEqual(1)
-        const firstError = allErrors[0]
-        expect(firstError.message).toMatch("Definition was declared but never used in document")
-        expect(firstError.level).toEqual("warning")
-        expect(firstError.path).toEqual(["definitions", "abc"])
-      })
-    })
-
-    it("should not return a warning when a definition with special character is declared and used", () => {
-      const spec = {
-        paths: {
-          "/CoolPath": {
-            get: {
-              responses: {
-                200: {
-                  schema: {
-                    $ref: "#/definitions/x~1Foo"
-                  }
-                },
-                400: {
-                  schema: {
-                    $ref: "#/definitions/x~0Bar"
-                  }
-                }
-              }
-            }
-          }
-        },
-        definitions: {
-          "x/Foo": {
-            type: "object"
-          },
-          "x~Bar": {
-            type: "object"
-          }
-        }
-      }
-
-      return validateHelper(spec)
-      .then(system => {
-        const allErrors = system.errSelectors.allErrors().toJS()
-        expect(allErrors.length).toEqual(0)
-      })
-    })
-
-  })
-  describe("Malformed $ref values", () => {
-    it("should return an error when a JSON pointer lacks a leading `#/`", () => {
-      const spec = {
-        paths: {
-          "/CoolPath": {
-            $ref: "#myObj/abc"
-          }
-        },
-        myObj: {
-          abc: {
-            type: "string"
-          }
-        }
-      }
-
-      return validateHelper(spec)
-      .then(system => {
-        const allErrors = system.errSelectors.allErrors().toJS()
-        expect(allErrors.length).toEqual(1)
-        const firstError = allErrors[0]
-        expect(firstError.message).toMatch("$ref paths must begin with `#/`")
-        expect(firstError.level).toEqual("error")
-        expect(firstError.path).toEqual(["paths", "/CoolPath", "$ref"])
-      })
-    })
-
-    it("should return no errors when a JSON pointer is a well-formed remote reference", () => {
-      const spec = {
-        paths: {
-          "/CoolPath": {
-            $ref: "http://google.com#/myObj/abc"
-          }
-        },
-        myObj: {
-          abc: {
-            type: "string"
-          }
-        }
-      }
-
-      return validateHelper(spec)
-      .then(system => {
-        const allSemanticErrors = system.errSelectors.allErrors().toJS()
-          .filter(err => err.source !== "resolver")
-        expect(allSemanticErrors).toEqual([])
-      })
-    })
-
-  })
   describe.skip("Refs are restricted in specific areas of a spec", () => {
     describe("Response $refs", () => {
       it("should return a problem for a parameters $ref in a response position", function(){
